Settings : Custom carrier text in notification pulldown
Allows changing the carrier text in the notification bar to a
customized string. A checkbox determines if the text will be displayed
and will clear out the setting when disabled. A reboot is needed for
the text to be applied or reset.

Originally written by Twisted
Rebased and fixed by me